Create README.md
You can create a digital clock with the help of Python after watching this Python alarm clock project. Learn how to make a GUI clock in python in just 10 lines of code.

You can do a lot of projects related to time in Python; here you will learn about making a digital clock using Tkinter. It is mainly used for UI in any Python project.

Tkinter is a Python package that deals with GUI part in it. After setting up the library and installing the required package, you can run the code to make a digital clock. Tk library is used in making the digital clock. 

You can select the dimensions and size of it as well. The size of the screen and font also can be changed in it. Python digital clock example will take you through this all.